Fixed a schwack load of issues.
- Swiping is working. SetDimension had been defined but didn't actually do a dimension = aDimension so it was always nil
- I've changed the collision to use masks and categories instead of groupIndicies this allows a cleaner impl for collision with the group and with the ground.
- Added a check for the swipe only on the frame it's detected. Swipe logic was firing every update.
- Updated the test level with a simple jumping puzzle.

+++ Projectfiles/Controller/Level.h
@@ -38,7 +38,7 @@
 
 - (b2Body *)initBody:(b2BodyDef *)fixture;
 
-- (void)initStaticBodies:(CCTMXTiledMap *)map collisionLayer:(NSString *)clLayer collisionGroupId:(int)collisionGroupId;
+- (void)initStaticBodies:(CCTMXTiledMap *)map collisionLayer:(NSString *)clLayer collisionGroupId:(uint16)collisionGroupId;
 
 - (void)spawn:(id <Updateable, Displayable>)objectToSpawn;
